% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_admin.R
\name{boxplot_proc}
\alias{boxplot_proc}
\title{Wykresy w raportach z 3. rundy monitoringu na danych administracyjnych}
\usage{
boxplot_proc(x, tytul)
}
\arguments{
\item{x}{ramka z danymi do przedstawienia na wykresie}

\item{tytul}{tytuł wykresu w formie ciągu tekstowego lub wartość NULL dla
braku tytułu}
}
\value{
wykres
}
\description{
Funkcja rysująca wykres typu boxplot na potrzeby raportu. Wykres
pokazuje odsetki.
}
