% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele_do_wykresow.R
\name{tab_facet}
\alias{tab_facet}
\title{Konwersja wskaźników w formie list na tabele do wsadu do wykresu}
\usage{
tab_facet(dane_szk, dane_god, dane_kraj, nazwa_god, typ_szk, szer = 80)
}
\arguments{
\item{dane_szk}{lista zawierająca wskaźniki na poziomie agregacji: szkoła
(lub wartość NULL jeśli ten poziom agregacji nie jest wymagany)}

\item{dane_god}{lista zawierająca wskaźniki na poziomie agregacji: grupa
odniesienia (lub wartość NULL jeśli ten poziom agregacji nie jest wymagany)}

\item{dane_kraj}{lista zawierająca wskaźniki na poziomie agregacji: cały kraj
(lub wartość NULL jeśli ten poziom agregacji nie jest wymagany)}

\item{nazwa_god}{nazwa poiatu, podregionu lub województwa}

\item{typ_szk}{typ szkoły, której dotyczy raport}

\item{szer}{szerokość tekstu - argument w funkcji \link{\code{wrapper}}}

\item{god}{nazwa poziomu podziału terytorialnego, w którym znajduje
się dana szkoła w formie dopełniacza (lub wartość NULL jeśli nie jest
wymagana)}
}
\value{
ramka danych w formacie tibble
}
\description{
Na potrzeby raportu funkcja tworzy z 3 list zawierających
wskaźniki na różnych poziomach agregacji: szkoła, województwo i cały kraj,
ramkę danych, która następnie jest używana jako wsad do funkcji rysującej
wyres \code{wyk_facet()}.
}
