% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_robocze.R
\name{tab_num}
\alias{tab_num}
\title{Funkcje robocze dla raportowania w 3. edycji monitoringu - dane
administracyjne}
\usage{
tab_num()
}
\value{
numeric
}
\description{
Funkcja zwracająca kolejne numery tabel (analogiczna do funkcji
\code{wyk_num}). Funkcja ta była częścią szablonów w poprzedniej edycji, a w
obecnej będzie częścią pakietu, żeby każdy raport nie ładował jej od nowa.
}
