% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_robocze.R
\name{wrapper}
\alias{wrapper}
\title{Funkcje robocze dla raportowania w 3. edycji monitoringu - dane
administracyjne}
\usage{
wrapper(x, ...)
}
\arguments{
\item{x}{ciąg znaków, który ma zostać podzielony}

\item{...}{dowolny argument funkcji \code{strwrap()} np. \code{width = 80}}
}
\value{
text
}
\description{
Funkcja przycinająca tekst do zadanej szerokości liczonej
długością znaków. Przy dzieleniu na wiersze całe słowa są brane pod uwagę.
Źródło: \link{https://stackoverflow.com/a/3935429}
}
\examples{
przykladowy_tekst = "Google trackers have been found on 75\% of the top million websites. This means they are not only tracking what you search for, they're also tracking which websites you visit, and using all your data for ads that follow you around the Internet. Your personal data can also be subpoenaed by lawyers, including for civil cases like divorce. Google answered over 150,000 such data requests in 2019 alone! More and more people are also realizing the risk of relying on one company for so many personal services."
wrapper(przykladowy_tekst, width = 80)
}
