% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_admin.R
\name{wyk_facet}
\alias{wyk_facet}
\title{Wykresy w raportach z 3. rundy monitoringu na danych administracyjnych}
\usage{
wyk_facet(x, tytul, os_x, tytul_x)
}
\arguments{
\item{x}{ramka z danymi do przedstawienia na wykresie}

\item{tytul}{tytuł wykresu w formie ciągu tekstowego lub wartość NULL dla
braku tytułu}

\item{os_x}{zmienna zawierająca etykiety osi x - miesiące lub przedziały}

\item{tytul_x}{tytuł osi x}
}
\value{
wykres
}
\description{
Funkcja rysująca wykres słupkowy, pionowy, skumulowany na
potrzeby raportu.
}
