% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_robocze.R
\name{wyk_num}
\alias{wyk_num}
\title{Funkcje robocze dla raportowania w 3. edycji monitoringu - dane
administracyjne}
\usage{
wyk_num()
}
\value{
numeric
}
\description{
Funkcja zwracająca kolejne numery wykresów. Funkcja ta była
częścią szablonów w poprzedniej edycji, a w obecnej będzie częścią pakietu,
żeby każdy raport nie ładował jej od nowa.
}
