% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_admin.R
\name{wykres_pion_skum}
\alias{wykres_pion_skum}
\title{Wykresy w raportach z 3. rundy monitoringu na danych administracyjnych}
\usage{
wykres_pion_skum(x, tytul)
}
\arguments{
\item{x}{ramka z danymi do przedstawienia na wykresie}

\item{tytul}{tytuł wykresu w formie ciągu tekstowego lub wartość NULL dla
braku tytułu}
}
\value{
wykres
}
\description{
Funkcja rysująca wykres słupkowy, pionowy, skumulowany na
potrzeby raportu.
}
