% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_admin.R
\name{wykres_poziomy_statusy}
\alias{wykres_poziomy_statusy}
\title{Wykresy w raportach z 3. rundy monitoringu na danych administracyjnych}
\usage{
wykres_poziomy_statusy(x, tytul)
}
\arguments{
\item{x}{ramka z danymi do przedstawienia na wykresie}

\item{tytul}{tytuł wykresu w formie ciągu tekstowego lub wartość NULL dla
braku tytułu}
}
\value{
wykres
}
\description{
Funkcja rysująca poziomy, skumulowany, wykres słupkowy na
potrzeby raportu. Jako argument zawierający zbiór danych należy podać ramkę
danych będącą wynikiem działania funkcji \code{tab_wykres_ad1()} lub
inną, analogiczną.
}
