% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_robocze.R
\name{zwroc_wykres_tabela}
\alias{zwroc_wykres_tabela}
\title{Funkcje robocze dla raportowania w 3. edycji monitoringu - dane
administracyjne}
\usage{
zwroc_wykres_tabela(wykres, tabela, tekst = TRUE)
}
\arguments{
\item{wykres}{obiekt klasy \code{ggplot}}

\item{tabela}{obiekt klasy \code{knitr_kable}}

\item{tekst}{wartość logiczna określająca czy ma być zwracany tekst "Dane z
wykresu przedstawiono w poniższej tabeli:"?}
}
\value{
lista
}
\description{
Funkcja używana w szablonie raportu, służąca do wyświetlania
wykresu oraz tabeli, w której pokazane są wartości z wykresu. Na wykresach
nie są prezentowane etykiety wartości mniejszych niż 6%, a z tabeli można je
odczytać.
}
